requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
mlflow
pandas
scikit-learn==1.3.2
pyyaml
great_expectations
psycopg2-binary
python-dotenv
joblib
xgboost
numpy
scipy
matplotlib
seaborn
jinja2
optuna
shap
imbalanced-learn==0.11.0
openpyxl
fastapi
uvicorn
streamlit
requests